refactored usb hid parser code (added modules), improved joystick support
