+++ src/usb_hid_keyboard.cpp
@@ -1,0 +1,211 @@
+#include "usb_hid_keyboard.h"
+
+#include <Arduino.h>
+#include <esp_log.h>
+#include "usb_hid_common.h"
+
+
+#include "hid_usage_keyboard.h"
+
+static const char* TAG = "usb-hid-keyboard";
+
+/**
+ * @brief Scancode to ascii table
+ */
+const uint8_t keycode2ascii[57][2] = {
+    {0, 0},     /* HID_KEY_NO_PRESS        */
+    {0, 0},     /* HID_KEY_ROLLOVER        */
+    {0, 0},     /* HID_KEY_POST_FAIL       */
+    {0, 0},     /* HID_KEY_ERROR_UNDEFINED */
+    {'a', 'A'}, /* HID_KEY_A               */
+    {'b', 'B'}, /* HID_KEY_B               */
+    {'c', 'C'}, /* HID_KEY_C               */
+    {'d', 'D'}, /* HID_KEY_D               */
+    {'e', 'E'}, /* HID_KEY_E               */
+    {'f', 'F'}, /* HID_KEY_F               */
+    {'g', 'G'}, /* HID_KEY_G               */
+    {'h', 'H'}, /* HID_KEY_H               */
+    {'i', 'I'}, /* HID_KEY_I               */
+    {'j', 'J'}, /* HID_KEY_J               */
+    {'k', 'K'}, /* HID_KEY_K               */
+    {'l', 'L'}, /* HID_KEY_L               */
+    {'m', 'M'}, /* HID_KEY_M               */
+    {'n', 'N'}, /* HID_KEY_N               */
+    {'o', 'O'}, /* HID_KEY_O               */
+    {'p', 'P'}, /* HID_KEY_P               */
+    {'q', 'Q'}, /* HID_KEY_Q               */
+    {'r', 'R'}, /* HID_KEY_R               */
+    {'s', 'S'}, /* HID_KEY_S               */
+    {'t', 'T'}, /* HID_KEY_T               */
+    {'u', 'U'}, /* HID_KEY_U               */
+    {'v', 'V'}, /* HID_KEY_V               */
+    {'w', 'W'}, /* HID_KEY_W               */
+    {'x', 'X'}, /* HID_KEY_X               */
+    {'y', 'Y'}, /* HID_KEY_Y               */
+    {'z', 'Z'}, /* HID_KEY_Z               */
+    {'1', '!'}, /* HID_KEY_1               */
+    {'2', '@'}, /* HID_KEY_2               */
+    {'3', '#'}, /* HID_KEY_3               */
+    {'4', '$'}, /* HID_KEY_4               */
+    {'5', '%'}, /* HID_KEY_5               */
+    {'6', '^'}, /* HID_KEY_6               */
+    {'7', '&'}, /* HID_KEY_7               */
+    {'8', '*'}, /* HID_KEY_8               */
+    {'9', '('}, /* HID_KEY_9               */
+    {'0', ')'}, /* HID_KEY_0               */
+    {KEYBOARD_ENTER_MAIN_CHAR, KEYBOARD_ENTER_MAIN_CHAR}, /* HID_KEY_ENTER */
+    {0, 0},      /* HID_KEY_ESC             */
+    {'\b', 0},   /* HID_KEY_DEL             */
+    {0, 0},      /* HID_KEY_TAB             */
+    {' ', ' '},  /* HID_KEY_SPACE           */
+    {'-', '_'},  /* HID_KEY_MINUS           */
+    {'=', '+'},  /* HID_KEY_EQUAL           */
+    {'[', '{'},  /* HID_KEY_OPEN_BRACKET    */
+    {']', '}'},  /* HID_KEY_CLOSE_BRACKET   */
+    {'\\', '|'}, /* HID_KEY_BACK_SLASH      */
+    {'\\', '|'},
+    /* HID_KEY_SHARP           */  // HOTFIX: for NonUS Keyboards repeat
+                                   // HID_KEY_BACK_SLASH
+    {';', ':'},                    /* HID_KEY_COLON           */
+    {'\'', '"'},                   /* HID_KEY_QUOTE           */
+    {'`', '~'},                    /* HID_KEY_TILDE           */
+    {',', '<'},                    /* HID_KEY_LESS            */
+    {'.', '>'},                    /* HID_KEY_GREATER         */
+    {'/', '?'}                     /* HID_KEY_SLASH           */
+};
+
+
+/**
+ * @brief HID Keyboard modifier verification for capitalization application
+ * (right or left shift)
+ *
+ * @param[in] modifier
+ * @return true  Modifier was pressed (left or right shift)
+ * @return false Modifier was not pressed (left or right shift)
+ *
+ */
+inline bool hid_keyboard_is_modifier_shift(uint8_t modifier) {
+    if (((modifier & HID_LEFT_SHIFT) == HID_LEFT_SHIFT) ||
+        ((modifier & HID_RIGHT_SHIFT) == HID_RIGHT_SHIFT)) {
+        return true;
+    }
+    return false;
+}
+
+/**
+ * @brief HID Keyboard get char symbol from key code
+ *
+ * @param[in] modifier  Keyboard modifier data
+ * @param[in] key_code  Keyboard key code
+ * @param[in] key_char  Pointer to key char data
+ *
+ * @return true  Key scancode converted successfully
+ * @return false Key scancode unknown
+ */
+inline bool hid_keyboard_get_char(uint8_t modifier, uint8_t key_code,
+                                         unsigned char* key_char) {
+    uint8_t mod = (hid_keyboard_is_modifier_shift(modifier)) ? 1 : 0;
+
+    if ((key_code >= HID_KEY_A) && (key_code <= HID_KEY_SLASH)) {
+        *key_char = keycode2ascii[key_code][mod];
+    } else {
+        // All other key pressed
+        return false;
+    }
+
+    return true;
+}
+
+/**
+ * @brief HID Keyboard print char symbol
+ *
+ * @param[in] key_char  Keyboard char to stdout
+ */
+inline void hid_keyboard_print_char(unsigned int key_char) {
+    if (!!key_char) {
+        putchar(key_char);
+#if (KEYBOARD_ENTER_LF_EXTEND)
+        if (KEYBOARD_ENTER_MAIN_CHAR == key_char) {
+            putchar('\n');
+        }
+#endif  // KEYBOARD_ENTER_LF_EXTEND
+        fflush(stdout);
+    }
+}
+
+/**
+ * @brief Key Event. Key event with the key code, state and modifier.
+ *
+ * @param[in] key_event Pointer to Key Event structure
+ *
+ */
+void key_event_callback(key_event_t* key_event) {
+    unsigned char key_char;
+
+    hid_print_new_device_report_header(HID_PROTOCOL_KEYBOARD);
+
+    if (key_event->KEY_STATE_PRESSED == key_event->state) {
+        if (hid_keyboard_get_char(key_event->modifier, key_event->key_code,
+                                  &key_char)) {
+            hid_keyboard_print_char(key_char);
+        }
+    }
+}
+
+/**
+ * @brief Key buffer scan code search.
+ *
+ * @param[in] src       Pointer to source buffer where to search
+ * @param[in] key       Key scancode to search
+ * @param[in] length    Size of the source buffer
+ */
+inline bool key_found(const uint8_t* const src, uint8_t key,
+                             unsigned int length) {
+    for (unsigned int i = 0; i < length; i++) {
+        if (src[i] == key) {
+            return true;
+        }
+    }
+    return false;
+}
+
+/**
+ * @brief USB HID Host Keyboard Interface report callback handler
+ *
+ * @param[in] data    Pointer to input report data buffer
+ * @param[in] length  Length of input report data buffer
+ */
+void hid_host_keyboard_report_callback(const uint8_t* const data, const int length) {
+    hid_keyboard_input_report_boot_t* kb_report =
+        (hid_keyboard_input_report_boot_t*)data;
+
+    if (length < sizeof(hid_keyboard_input_report_boot_t)) {
+        return;
+    }
+
+    static uint8_t prev_keys[HID_KEYBOARD_KEY_MAX] = {0};
+    key_event_t key_event;
+
+    for (int i = 0; i < HID_KEYBOARD_KEY_MAX; i++) {
+        // key has been released verification
+        if (prev_keys[i] > HID_KEY_ERROR_UNDEFINED &&
+            !key_found(kb_report->key, prev_keys[i], HID_KEYBOARD_KEY_MAX)) {
+            key_event.key_code = prev_keys[i];
+            key_event.modifier = 0;
+            key_event.state = key_event.KEY_STATE_RELEASED;
+            key_event_callback(&key_event);
+        }
+
+        // key has been pressed verification
+        if (kb_report->key[i] > HID_KEY_ERROR_UNDEFINED &&
+            !key_found(prev_keys, kb_report->key[i], HID_KEYBOARD_KEY_MAX)) {
+            key_event.key_code = kb_report->key[i];
+            key_event.modifier = kb_report->modifier.val;
+            key_event.state = key_event.KEY_STATE_PRESSED;
+            key_event_callback(&key_event);
+        }
+    }
+
+    memcpy(prev_keys, &kb_report->key, HID_KEYBOARD_KEY_MAX);
+}
+
