+++ src/usb_hid_keyboard.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "hid_host.h"
+
+
+/**
+ * @brief Key event
+ */
+typedef struct {
+    enum key_state {
+        KEY_STATE_PRESSED = 0x00,
+        KEY_STATE_RELEASED = 0x01
+    } state;
+    uint8_t modifier;
+    uint8_t key_code;
+} key_event_t;
+
+/* Main char symbol for ENTER key */
+#define KEYBOARD_ENTER_MAIN_CHAR '\r'
+/* When set to 1 pressing ENTER will be extending with LineFeed during serial
+ * debug output */
+#define KEYBOARD_ENTER_LF_EXTEND 1
+
+
+void hid_host_keyboard_report_callback(const uint8_t* const data, const int length);
